@@ -28,7 +28,7 @@ public class AppRunner {
       @Override
       public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
-            .allowedMethods("GET", "POST")
+            .allowedMethods("GET", "POST", "PUT", "DELETE")
             .allowedOrigins("*");
       }
     };
