@@ -5,9 +5,9 @@ package io.catalyte.training.sportsproducts.domains.user;
  */
 public interface UserService {
 
-  public User updateUser(String credentials, Long id, User user);
+  User updateUser(String credentials, Long id, User user);
 
-  public User createUser(User user);
+  User createUser(User user);
 
-  public User getUserByEmail(String email);
+  User getUserByEmail(String email);
 }
