@@ -13,7 +13,9 @@ import io.catalyte.training.sportsproducts.data.ProductFactory;
 import io.catalyte.training.sportsproducts.exceptions.DataAccessError;
 import io.catalyte.training.sportsproducts.exceptions.ResourceNotFound;
 import io.catalyte.training.sportsproducts.exceptions.ServerError;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import org.junit.Before;
@@ -27,6 +29,9 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 
+/**
+ * Unit testing for Product
+ */
 @RunWith(MockitoJUnitRunner.class)
 @WebMvcTest(ProductServiceImpl.class)
 public class ProductServiceImplTest {
@@ -70,6 +75,8 @@ public class ProductServiceImplTest {
 
   /**
    * Get a product by ID from the productRepository and assert that it has the required fields
+   *
+   * @author - Andrew Salerno
    */
   @Test
   public void getProductByIdHasRequiredFields() {
@@ -97,9 +104,11 @@ public class ProductServiceImplTest {
 
   /**
    * getProducts() data access exception test
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void getProductsReturnsDataAccessExceptionMessage () {
+  public void getProductsReturnsDataAccessExceptionMessage() {
     when(productRepository.findAll()).thenThrow(new DataAccessError(""));
     try {
       Map<String, String> allParams = new HashMap<>();
@@ -113,9 +122,11 @@ public class ProductServiceImplTest {
 
   /**
    * getProductById() data access exception test
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void getProductByIdReturnsDataAccessExceptionMessage () {
+  public void getProductByIdReturnsDataAccessExceptionMessage() {
     when(productRepository.findById(anyLong())).thenThrow(
         new DataAccessError(""));
     try {
@@ -126,14 +137,34 @@ public class ProductServiceImplTest {
       String expectedServerMessage = "";
       assertEquals(expectedServerMessage, actualServerMessage);
     }
+  }
 
+  /**
+   * saveProduct() data access exception test
+   *
+   * @author - Andrew Salerno
+   */
+  @Test
+  public void saveProductReturnsDataAccessExceptionMessage() {
+    Product product = new Product();
+    when(productRepository.save(product)).thenThrow(
+        new DataAccessError(""));
+    try {
+      productServiceImpl.saveProduct(product);
+    } catch (ServerError e) {
+      String actualServerMessage = e.getMessage();
+      String expectedServerMessage = "";
+      assertEquals(expectedServerMessage, actualServerMessage);
+    }
   }
 
   /**
    * getUniqueTypes() data access exception test
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void getUniqueTypesReturnsDataAccessExceptionMessage () {
+  public void getUniqueTypesReturnsDataAccessExceptionMessage() {
     when(productRepository.findByType()).thenThrow(new DataAccessError(""));
     try {
       productServiceImpl.getUniqueTypes();
@@ -145,10 +176,29 @@ public class ProductServiceImplTest {
   }
 
   /**
-   * getUniqueCategories() data access exception test
+   * getUniqueDemographics() data access exception test
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void getUniqueCategoriesReturnsDataAccessExceptionMessage () {
+  public void getUniqueDemographicsReturnsDataAccessExceptionMessage() {
+    when(productRepository.findByDemographic()).thenThrow(new DataAccessError(""));
+    try {
+      productServiceImpl.getUniqueDemographics();
+    } catch (ServerError e) {
+      String actualServerMessage = e.getMessage();
+      String expectedServerMessage = "";
+      assertEquals(expectedServerMessage, actualServerMessage);
+    }
+  }
+
+  /**
+   * getUniqueCategories() data access exception test
+   *
+   * @author - Andrew Salerno
+   */
+  @Test
+  public void getUniqueCategoriesReturnsDataAccessExceptionMessage() {
     when(productRepository.findByCategory()).thenThrow(new DataAccessError(""));
     try {
       productServiceImpl.getUniqueCategories();
@@ -160,10 +210,63 @@ public class ProductServiceImplTest {
   }
 
   /**
-   * Checks product.toString() returns the correct format
+   * getUniqueBrands() data access exception test
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void productToStringReturnsCorrectFormat () {
+  public void getUniqueBrandsReturnsDataAccessExceptionMessage() {
+    when(productRepository.findByBrand()).thenThrow(new DataAccessError(""));
+    try {
+      productServiceImpl.getUniqueBrands();
+    } catch (ServerError e) {
+      String actualServerMessage = e.getMessage();
+      String expectedServerMessage = "";
+      assertEquals(expectedServerMessage, actualServerMessage);
+    }
+  }
+
+  /**
+   * getUniqueMaterials() data access exception test
+   *
+   * @author - Andrew Salerno
+   */
+  @Test
+  public void getUniqueMaterialsReturnsDataAccessExceptionMessage() {
+    when(productRepository.findByMaterial()).thenThrow(new DataAccessError(""));
+    try {
+      productServiceImpl.getUniqueMaterials();
+    } catch (ServerError e) {
+      String actualServerMessage = e.getMessage();
+      String expectedServerMessage = "";
+      assertEquals(expectedServerMessage, actualServerMessage);
+    }
+  }
+
+  /**
+   * getUniqueColors() data access exception test
+   *
+   * @author - Andrew Salerno
+   */
+  @Test
+  public void getUniqueColorsReturnsDataAccessExceptionMessage() {
+    when(productRepository.findByPrimaryColor()).thenThrow(new DataAccessError(""));
+    try {
+      productServiceImpl.getUniqueColors();
+    } catch (ServerError e) {
+      String actualServerMessage = e.getMessage();
+      String expectedServerMessage = "";
+      assertEquals(expectedServerMessage, actualServerMessage);
+    }
+  }
+
+  /**
+   * Checks product.toString() returns the correct format
+   *
+   * @author - Andrew Salerno
+   */
+  @Test
+  public void productToStringReturnsCorrectFormat() {
     Product product = new Product();
     String actual = product.toString();
     String expected =
@@ -179,9 +282,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.hashCode() returns
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void productHashCodeReturnsResult () {
+  public void productHashCodeReturnsResult() {
     Product product = new Product();
     int actual = product.hashCode();
     int expected = product.hashCode();
@@ -190,9 +295,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false after passing a Null object
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfProductIsNull () {
+  public void equalsReturnsFalseIfProductIsNull() {
     Product actual = new Product();
     Product expected = new Product();
     expected = null;
@@ -200,11 +307,13 @@ public class ProductServiceImplTest {
   }
 
   /**
-   * Creates a Product with the Product constructor and then sets its id with product.setId()
-   * then gets the product id with product.getId() and then compares the values
+   * Creates a Product with the Product constructor and then sets its id with product.setId() then
+   * gets the product id with product.getId() and then compares the values
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void constructProductSetIDGetID () {
+  public void constructProductSetIDGetID() {
     Product product = new Product("name", "description", "demographic",
         "category", "type", "releaseDate");
     long expected = 9999L;
@@ -216,9 +325,11 @@ public class ProductServiceImplTest {
 
   /**
    * Creates a Product with the Product constructor, gets its name and compares
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void constructProductAndGetName () {
+  public void constructProductAndGetName() {
     Product product = new Product("name", "description", "demographic",
         "category", "type", "releaseDate");
     String expected = "name";
@@ -230,9 +341,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same name
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentName () {
+  public void equalsReturnsFalseIfObjectHasDifferentName() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setName("name");
@@ -243,9 +356,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same brand
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentBrand () {
+  public void equalsReturnsFalseIfObjectHasDifferentBrand() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setBrand("brand");
@@ -256,9 +371,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same description
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentDescription () {
+  public void equalsReturnsFalseIfObjectHasDifferentDescription() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setDescription("description");
@@ -269,9 +386,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same demographic
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentDemographic () {
+  public void equalsReturnsFalseIfObjectHasDifferentDemographic() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setDemographic("demographic");
@@ -282,9 +401,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same category
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentCategory () {
+  public void equalsReturnsFalseIfObjectHasDifferentCategory() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setCategory("category");
@@ -295,9 +416,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same type
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentType () {
+  public void equalsReturnsFalseIfObjectHasDifferentType() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setType("type");
@@ -308,9 +431,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same releaseDate
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentReleaseDate () {
+  public void equalsReturnsFalseIfObjectHasDifferentReleaseDate() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setReleaseDate("releaseDate");
@@ -321,10 +446,12 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same price
+   *
+   * @author - Andrew Salerno
    */
   @Test
 
-  public void equalsReturnsFalseIfObjectHasDifferentPrice () {
+  public void equalsReturnsFalseIfObjectHasDifferentPrice() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setPrice(1.00);
@@ -335,9 +462,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same primaryColorCode
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentPrimaryColorCode () {
+  public void equalsReturnsFalseIfObjectHasDifferentPrimaryColorCode() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setPrimaryColorCode("primaryColorCode");
@@ -348,9 +477,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same quantity
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentQuantity () {
+  public void equalsReturnsFalseIfObjectHasDifferentQuantity() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setQuantity(1);
@@ -361,9 +492,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same secondaryColorCode
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentSecondaryColorCode () {
+  public void equalsReturnsFalseIfObjectHasDifferentSecondaryColorCode() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setSecondaryColorCode("secondaryColorCode");
@@ -374,9 +507,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same styleNumber
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentStyleNumber () {
+  public void equalsReturnsFalseIfObjectHasDifferentStyleNumber() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setStyleNumber("styleNumber");
@@ -387,9 +522,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same globalProductCode
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentGlobalProductCode () {
+  public void equalsReturnsFalseIfObjectHasDifferentGlobalProductCode() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setGlobalProductCode("globalProductCode");
@@ -400,9 +537,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same imageSrc
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentImageSrc () {
+  public void equalsReturnsFalseIfObjectHasDifferentImageSrc() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setImageSrc("imageSrc");
@@ -413,9 +552,11 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns false if Products have the same material
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsFalseIfObjectHasDifferentMaterial () {
+  public void equalsReturnsFalseIfObjectHasDifferentMaterial() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setMaterial("material");
@@ -426,14 +567,68 @@ public class ProductServiceImplTest {
 
   /**
    * Checks that product.equals() returns an active product
+   *
+   * @author - Andrew Salerno
    */
   @Test
-  public void equalsReturnsTrueIfProductIsActive () {
+  public void equalsReturnsTrueIfProductIsActive() {
     Product actual = new Product();
     Product expected = new Product();
     actual.setActive(true);
     expected.setActive(false);
     boolean result = actual.equals(expected);
     assertFalse(result, "Products have the same active status");
+  }
+
+  /**
+   * Checks that an empty list is returned when empty parameters are passed to getProducts()
+   *
+   * @author - AndrewSalerno
+   */
+  @Test
+  public void getProductsReturnsEmptyWithEmptyParams() {
+    Map<String, String> allParams = new HashMap<>();
+    allParams.put("", "");
+    List<Product> actual = productServiceImpl.getProducts(allParams);
+    List<Product> expected = new ArrayList<>();
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Checks that getProducts() returns and empty list when invalid parameters are passed
+   *
+   * @author - AndrewSalerno
+   */
+  @Test
+  public void getProductsReturnsEmptyOnInvalidParams() {
+    Map<String, String> allParams = new HashMap<>();
+    allParams.put("NotValidKey", "NotValidValue");
+    List<Product> actual = productServiceImpl.getProducts(allParams);
+    List<Product> expected = new ArrayList<>();
+    assertEquals(expected, actual);
+  }
+
+  /**
+   * Checks that getProducts() returns a list of products when valid parameters are passed
+   *
+   * @author - Andrew Salerno
+   */
+  @Test
+  public void getProductsReturnsProductsOnValidParams() {
+    Map<String, String> allParams = new HashMap<>();
+    allParams.put("price", "100.00");
+    List<Product> actual = productServiceImpl.getProducts(allParams);
+    List<Product> expected = new ArrayList<>();
+    assertEquals(expected, actual);
+  }
+
+
+  @Test
+  public void saveProductReturnsNewProduct() {
+    Product expected = new Product();
+    expected.setId(999L);
+
+    Product actual = productServiceImpl.saveProduct(expected);
+    assertEquals(expected, actual);
   }
 }
